requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
et-xmlfile==1.1.0
openpyxl==3.0.7
Pillow==8.2.0
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
